+++ package.json
@@ -27,6 +27,7 @@
     "ember-cli-inject-live-reload": "^1.3.0",
     "ember-cli-less": "1.3.3",
     "ember-cli-qunit": "0.3.13",
+    "ember-cli-htmlbars-inline-precompile": "0.3.1",
     "ember-cli-uglify": "^1.0.1",
     "ember-disable-prototype-extensions": "^1.0.0",
     "ember-disable-proxy-controllers": "^1.0.0",
