+++ tests/test-helper.js
@@ -3,10 +3,10 @@ import {
   setResolver
 } from 'ember-qunit';
 
-import './helpers/assertions';
-import './helpers/mouse';
-import './helpers/keyboard';
-import './helpers/color-picker';
-import './helpers/select';
+// import './helpers/assertions';
+// import './helpers/mouse';
+// import './helpers/keyboard';
+// import './helpers/color-picker';
+// import './helpers/select';
 
 setResolver(resolver);
